# backend/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'VanScape API'
  description: 'Vanscape API endpoints'
  version: 1.0.0
servers:
  -
    url: 'https://api-vanscape.duckdns.org'
    description: 'API over HTTPS'
paths:
...
  /register:
    post:
      tags:
        - User
      summary: 'Create new user.'
      operationId: post_create_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: 'Returns data for the created user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO4'
        '400':
          description: 'Errors found in received data in %s'
        '409':
          description: 'User already exists'
        '500':
          description: 'Oops ! Something went wrong.'
  /api/users:
    get:
      tags:
        - User
      summary: 'Retrieves the authenticated user informations.'
      operationId: get_read_user
      responses:
        '200':
          description: 'Returns data for a user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO4'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'User not found'
        '500':
          description: 'Oops ! Something went wrong.'
    put:
      tags:
        - User
      summary: 'Update authenticated user informations.'
      operationId: put_edit_user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO2'
      responses:
        '200':
          description: 'Returns updated data for the authenticated user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO4'
        '400':
          description: 'Errors found in received data in %s'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'User not found'
        '409':
          description: 'User already exists'
        '500':
          description: 'Oops ! Something went wrong.'
    delete:
      tags:
        - User
      summary: 'Deletes the authenticated user account.'
      operationId: delete_delete_user
      responses:
        '204':
          description: 'Successfully deleted the user'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'User not found'
        '500':
          description: 'Oops ! Something went wrong.'
    patch:
      tags:
        - User
      summary: 'Update authenticated user password.'
      operationId: patch_edit_user_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPasswordDTO'
      responses:
        '204':
          description: 'Successfully updated the password'
        '400':
          description: 'Errors found in received data in %s'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'User not found'
        '403':
          description: 'You are not authorized to perform this action (user)'
        '500':
          description: 'Oops ! Something went wrong.'
  /api/users/summary:
    get:
      tags:
        - User
      summary: 'Retrieves user extra infos summary (number of spots and friends).'
      operationId: get_user_extra_infos
      responses:
        '200':
          description: 'Returns extra informations data for the authenticated user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummaryDTO'
        '401':
          description: 'No authenticated user found in JWT'
        '404':
          description: 'User not found'
        '500':
          description: 'Oops ! Something went wrong.'
components:
  schemas:
...
    UserDTO:
      title: User
      description: 'User identity'
      required:
        - email
        - pseudo
      properties:
        email:
          description: 'Email of the user'
          type: string
          example: jane.doe@gmail.com
        pseudo:
          description: 'Pseudo of the user'
          type: string
          maxLength: 50
          minLength: 3
          pattern: '[a-zA-Z0-9_]+'
          example: jane_doe
        emailVerified:
          description: 'Not implemented yet'
          type: boolean
          default: false
          nullable: true
        password:
          description: 'Hashed password'
          type: string
          default: null
          minLength: 8
          nullable: true
      type: object
    UserDTO2:
      title: User
      description: 'User identity'
      required:
        - email
        - pseudo
      properties:
        email:
          description: 'Email of the user'
          type: string
          example: jane.doe@gmail.com
        pseudo:
          description: 'Pseudo of the user'
          type: string
          maxLength: 50
          minLength: 3
          pattern: '[a-zA-Z0-9_]+'
          example: jane_doe
        picture:
          description: 'Choosen avatar for profil picture'
          type: string
          nullable: true
        emailVerified:
          description: 'Not implemented yet'
          type: boolean
          default: false
          nullable: true
      type: object
    UserPasswordDTO:
      title: 'User password'
      description: 'Update password DTO'
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          description: 'Plain current password'
          type: string
          minLength: 8
        newPassword:
          description: 'Plain new password'
          type: string
          minLength: 8
      type: object
...
    UserDTO4:
      title: User
      description: 'User identity'
      required:
        - id
        - email
        - pseudo
        - createdAt
        - updatedAt
      properties:
        id:
          description: 'User id'
          type: integer
          example: 12
        email:
          description: 'Email of the user'
          type: string
          example: jane.doe@gmail.com
        pseudo:
          description: 'Pseudo of the user'
          type: string
          maxLength: 50
          minLength: 3
          pattern: '[a-zA-Z0-9_]+'
          example: jane_doe
        createdAt:
          description: 'Creation date of the account'
          type: string
          format: date-time
        updatedAt:
          description: 'Updated date of the account'
          type: string
          format: date-time
        picture:
          description: 'Choosen avatar for profil picture'
          type: string
          nullable: true
        token:
          description: 'JWT token for a session'
          type: string
          nullable: true
        emailVerified:
          description: 'Not implemented yet'
          type: boolean
          default: false
          nullable: true
      type: object
    UserSummaryDTO:
      title: 'User summary'
      description: 'Extra infos about user'
      required:
        - friendsNumber
        - spotsNumber
      properties:
        friendsNumber:
          description: 'Number of friends'
          type: integer
        spotsNumber:
          description: 'Number of spots'
          type: integer
      type: object
...
  securitySchemes:
    Bearer:
      type: http
      bearerFormat: JWT
      scheme: bearer
security:
  -
    Bearer: []
...